bit more of an object oriented approach with db connection
